Refactor text resource handling
