+++ OTAnalytics/plugin_ui/customtkinter_gui/frame_files.py
@@ -4,13 +4,11 @@ from typing import Any
 
 from customtkinter import CTkButton, CTkFrame, CTkScrollbar
 
+from OTAnalytics.adapter_ui.text_resources import ColumnResource
 from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.plugin_ui.customtkinter_gui.constants import PADX, PADY, STICKY
 from OTAnalytics.plugin_ui.customtkinter_gui.custom_containers import EmbeddedCTkFrame
-from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import (
-    ColumnResource,
-    TreeviewTemplate,
-)
+from OTAnalytics.plugin_ui.customtkinter_gui.treeview_template import TreeviewTemplate
 
 
 class FrameFiles(EmbeddedCTkFrame):
